fix: fixed id typing on dto request

+++ src/main/java/com/challenge/geosapiens/service_a/infrastructure/dto/response/OrderResponse.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class OrderResponse {
     private Long id;
     private String description;
-    private Integer userId;
-    private Integer deliveryPersonId;
+    private Long userId;
+    private Long deliveryPersonId;
 }
